feat: Start activities on the phone via the wear app!!
Finally, a full cycle, MVP of the wear app!

This commit adds the ability to start activities
on the phone from the wear app. Choose an
activity and the associated tag, and the phone app
will update accordingly.

Of course, the app still has plenty of bugs and
rough edges, but this is a huge milestone!

closes thehale/SimpleTimeTracker-WearOS#5
closes thehale/SimpleTimeTracker-WearOS#8