feat(wear): Show tag colors in TagList
Requires pulling tag colors from the phone (which
itself requires reconciling the color based on
whether or not the tag is a general tag), then
rendering them in the tag chip.

Also took the liberty to switch the tag chip from
a ToggleChip base to a SplitToggleChip base, which
(1) looks better (i.e. more consistent with the
ActivitiesList) and (2) allows multi-taggers to
start an activity immediately upon selecting one
tag, if desired.